refactor: global header component
