+++ webapp/src/components/header/header.module.scss
@@ -1,4 +1,4 @@
-@use "../styles/variables/sass.scss" as *;
+@use "../../styles/variables/sass.scss" as *;
 
 .block {
   display: flex;
